requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
openai
python-dotenv
pydantic
faiss-cpu
tiktoken
pdfplumber
pytesseract
pillow
pytest
python-multipart
